Resarch-Assistant-Agent/requirements.txt
langchain
langchain-groq
langchain-community
langchain-pinecone
langgraph
pinecone-client
sentence-transformers
newsapi-python
requests
python-dotenv
fastapi
uvicorn
httpx
tavily_python